adiciona verificação de autor no caso de uso de deletar perguntas

+++ src/domain/forum/application/use-cases/delete-question.ts
@@ -4,6 +4,7 @@ import type { QuestionsRepository } from '../repositories/questions-repository'
 
 interface DeleteQuestionUseCaseRequest {
   questionId: string
+  authorId: string
 }
 
 // biome-ignore lint/suspicious/noEmptyInterface: <explanation>
@@ -14,10 +15,14 @@ export class DeleteQuestionUseCase {
 
   async execute({
     questionId,
+    authorId,
   }: DeleteQuestionUseCaseRequest): Promise<DeleteQuestionUseCaseResponse> {
     const question = await this.questionsRepository.findById(questionId)
 
     if (!question) throw new Error('Question not found!')
+
+    if (authorId !== question.authorId.toString())
+      throw new Error('Not allowerd')
 
     await this.questionsRepository.delete(question)
